==> Lessons/03_DataExploration_Part2.Rmd <==
---
title: "5: Data Exploration"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2022"
output: pdf_document
geometry: margin=2.54cm
fig_width: 5
fig_height: 2.5
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Import and explore datasets in R
2. Graphically explore datasets in R
3. Apply data exploration skills to a real-world example dataset

## Opening discussion: why do we explore our data?

Why is data exploration our first step in analyzing a dataset? What information do we gain? How does data exploration aid in our decision-making for data analysis steps further down the pipeline?

## Import data and view summaries

```{r, message = FALSE}
# 1. Set up your working directory
getwd()

# 2. Load packages
library(tidyverse)

# 3. Import datasets
USGS.flow.data <- read.csv("./Data/Processed/USGS_Site02085000_Flow_Processed.csv")

#View(USGS.flow.data)
# Alternate option: click on data frame in Environment tab

colnames(USGS.flow.data)
str(USGS.flow.data)
dim(USGS.flow.data)

# Check our date column
class(USGS.flow.data$datetime)

USGS.flow.data$datetime <- as.Date(USGS.flow.data$datetime, format = "%Y-%m-%d") 
class(USGS.flow.data$datetime)
##CSV DATES ARE ALWAYS chr BY DEFAULT, MUST BE CHANGES TO Date OBJECT^

```


## Visualization for Data Exploration

Although the `summary()` function is helpful in getting an idea of the spread of values in a numeric dataset, it can be useful to create visual representations of the data to help form hypotheses and direct downstream data analysis. Below is a summary of the useful types of graphs for data exploration.

Note: each of these approaches utilize the package "ggplot2". We will be covering the syntax of ggplot in a later lesson, but for now you should familiarize yourself with the functionality of what each command is doing.

### Bar Chart (function: geom_bar)

Visualize count data for categorical variables. 

```{r, fig.height = 3, fig.width = 4}
ggplot(USGS.flow.data, aes(x = discharge.mean.approval)) +
  geom_bar()
```

### Histogram (function: geom_histogram)

Visualize distributions of values for continuous numerical variables. What is happening in each line of code? Insert a comment above each line.

```{r, fig.height = 3, fig.width = 4}
#plots histogram of discharge.mean with default visualization parameters
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean))

#plots histogram of discharge.mean with bindwidth of 10 (10 samples/bin)
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean), binwidth = 10)

#plots histogram of discharge.mean with 20 bins (NOTICE this looks like the 
#first example, but they are different)
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean), bins = 20)

#plots histogram of discharge.mean with binwidth of 10, & sample below 500 mean
# discharge volume
ggplot(USGS.flow.data, aes(x = discharge.mean)) +
  geom_histogram(binwidth = 10) + 
  scale_x_continuous(limits = c(0, 500))
  
#plots histogram of gage.height.mean with default visualization parameters
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = gage.height.mean))

```
### Frequency line graph (function: geom_freqpoly)

An alternate to a histogram is a frequency polygon graph (distributions of values for continuous numerical variables). Instead of displaying bars,  counts of continuous variables are displayed as lines. This is advantageous if you want to display multiple variables or categories of variables at once.

```{r, fig.height = 3, fig.width = 4}
#plots frequency lines (# of samples) for gage.height.mean, .max, & .mean each
#not ways for distinguishing variables (color, lty -> 'line type')
ggplot(USGS.flow.data) +
  geom_freqpoly(aes(x = gage.height.mean), bins = 50) +
  geom_freqpoly(aes(x = gage.height.min), bins = 50, color = "red") +
  geom_freqpoly(aes(x = gage.height.max), bins = 50,  lty = 2) +
  scale_x_continuous(limits = c(0, 10))

#plots frequency lines for gage.height.mean, but seperates and symbolizes fre-
#quencies by another variable, in this case gage.height.mean.approval
ggplot(USGS.flow.data) +
  geom_freqpoly(aes(x = gage.height.mean, color = gage.height.mean.approval), bins = 50) +
  scale_x_continuous(limits = c(0, 10)) +
  theme(legend.position = "top")

```
### Box-and-whisker plots (function: geom_boxplot, geom_violin)

A box-and-whisker plot is yet another alternative to histograms (distributions of values for continuous numerical variables). These plots consist of: 

* A box from the 25th to the 75th percentile of the data, called the interquartile range (IQR).

* A bold line inside the box representing the median value of the data. Whether the median is in the center or off to one side of the IQR will give you an idea about the skewness of your data.

* A line outside of the box representing values falling within 1.5 times the IQR. 

* Points representing outliers, values that fall outside 1.5 times the IQR. 

An alternate option is a violin plot, which displays density distributions, somewhat like a hybrid of the box-and-whiskers and the frequency polygon plot.

```{r, fig.height = 3, fig.width = 4}
#creates boxplots of gage.height.mean for unique values in gage.height.mean.approval
ggplot(USGS.flow.data) +
  geom_boxplot(aes(x = gage.height.mean.approval, y = gage.height.mean))

#creates boxplots of discharge.mean by gage.height.mean within each gage.height.
#mean ("cut_width= 1")
ggplot(USGS.flow.data) +
  geom_boxplot(aes(x = gage.height.mean, y = discharge.mean, group = cut_width(gage.height.mean, 1)))

#creates violin plot of same variables as first example
ggplot(USGS.flow.data) +
  geom_violin(aes(x = gage.height.mean.approval, y = gage.height.mean), 
              draw_quantiles = c(0.25, 0.5, 0.75))
```

### Scatterplot (function: geom_point)
Visualize relationships between continuous numerical variables.

```{r, fig.height = 3, fig.width = 4}
ggplot(USGS.flow.data) +
  geom_point(aes(x = discharge.max, y = gage.height.max))

ggplot(USGS.flow.data) +
  geom_point(aes(x = datetime, y = discharge.mean))
```

Question: under what circumstances would it be beneficial to use each of these graph types (bar plot, histogram, frequency polygon, box-and whisker, violin, scatterplot)?

> Answer: It really depends on what you're interested in. Some tell you about the frequency and distribution of data (bar plot, histogram, frequency polygon, violin), others give a lite statistical summary (box-and-whisker), and others let you look for any trends between different variables(scatterplot). 

## Ending discussion

What did you learn about the USGS discharge dataset today? What separate insights did the different graph types offer?
> Answer: The thing I learned that stands out most to me is that the gage site this dataset came from experienced a gap in monitoring between the early 1970's and mid 1980's. The bar plot and histogram showed the frequency and distribution of values for a single variable. The frequency polygon did the same but allowed us to compare multiple variables at the same time. The box-and-whisker plot showed us, very generally, the tendencies within the data (by which I mean how the data may be skewed & where the median is). The violin plot did the same while also visualizing the frequency of datapoints sharing the same value. The scatterplot allowed us to observe relationships between variables and potentially identify correlation or significant relationships.


How can multiple options for data exploration inform our understanding of our data?

> Answer: Since each type of visualization tells us something slightly different, using many different types before beginning any analyses could reveal patterns we hadn't expected before. Thus offering additional avenues for investigation. 


Do you see any patterns in the USGS data for the Eno River? What might be responsible for those patterns and/or relationships?

> Answer: I saw the monitoring gap I mentioned earlier. Possible explanations for this include lost data, a gap in program support ($$$), or some other explanation (perhaps development was occuring on or around the river so much that monitoring was suspended so the otherwise 'normal' data set wouldn't be confounded with atypical flow patters; no idea why anyone would do that though). I also noticed (after playing around a bit) that gage.height.max does not really increase after a certain discharge.mean. In my mind this means that perhaps the gage was overwhelmed? I also plotted gage.height.max and discharge.max, but found no similair pattern, so I can't say anything with certainty.
